% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Bin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of success}
}
\value{
Creates a framework to run binomial distributions
}
\description{
My Bin
}
\examples{
\dontrun{mybin(iter=1000,n=18, p=0.3)}

}
