% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot 2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample vector}

\item{fun}{function of interest}

\item{alpha}{confidence measure}

\item{cx}{plot label text size}

\item{...}{Additional arguments (not used in this function)}
}
\value{
If the sample is of size n then the bootstrap re-sampling will create
samples of size n with replacement. The following is a function that will
facilitate the bootstrap procedure.
}
\description{
My Boot 2
}
\examples{
\dontrun{myboot2(iter=10000,x=sam,fun="mean",alpha=0.05,cx=1.5,...)}


}
