% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My CLT}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}
}
\value{
Creates a uniform distribution of the central limit theorem, given a
set sample size and number of iterations.
}
\description{
My CLT
}
\examples{
\dontrun{myclt(n=50,iter=10000)}

}
