% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My NCurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard Deviation}

\item{a}{Upper Limit of Integration}
}
\value{
displays the curve, the shaded area between the curve, and the
x-axis from -∞ to x=a, and calculate the area (probability, P(X<=a)) which is
released to the command-line in a list.
}
\description{
My NCurve
}
\examples{
\dontrun{myncurve(mu=10,sigma=5, a=3)}

}
