% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats available on a flight}

\item{gamma}{acceptable risk (or probability) of overbooking a flight}

\item{p}{probability of a customer that buys a ticket actually showing up
to the flight}
}
\value{
This function finds the optimal number of tickets for an airline to
sell in order to have a full flight (N seats filled), taking into account
gamma risk of accidentally overbooking and p probability of a customer who
buys a ticket actually showing up to the flight.
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=400,gamma=0.02,p=0.95)}


}
